Implementando lógica da prova no front
