+++ bu_frontend/src/views/pages/Monitorar/logic/client.js
@@ -10,13 +10,14 @@ client.on('connect', function () {
       console.log("Conectado")
   })
 })
-const clientmqttSubConsistency = mqtt.connect(mosquitto_url)
+/* ----------------------------------------------------------------- */
+const clientConsistency = mqtt.connect(mosquitto_url)
 
-clientmqttSubConsistency.on('connect', function () {
-  clientmqttSubConsistency.subscribe('guilherme/consistencyProof', {qos: 2}, function (err) {
+clientConsistency.on('connect', function () {
+  clientConsistency.subscribe('guilherme/consistencyProof', {qos: 2}, function (err) {
     if (!err) 
       console.log("Conectado")
   })
 })
 
-export default (client,clientmqttSubConsistency)
+export default (client,clientConsistency)
